requirements.txt
asgiref==3.4.1
Django==3.2.5
django-admin-honeypot==1.1.0
django-environ==0.4.5
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.3.1
psycopg2-binary==2.9.1
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
